NEON tbl compress kernels, and measured density bands for x86
Adds a NEON tier to the fixed-width filter compress kernels and retunes the
x86 tier against real AVX-512 hardware.

NEON has no `vpcompress`, but `tbl` *is* a cross-lane byte shuffle, so a
byte-index LUT covering one vector of lanes serves every element width -
unlike AVX2, which has no byte shuffle and so leaves 1/2-byte elements on the
scalar byte-LUT. Lanes per shuffle is 8/8/4/2 for 1/2/4/8-byte elements.

The chunk loop is deliberately branchless. An `m != 0` guard looks like a
cheap skip but is a large pessimization: the mask bits of one chunk are as
random as the data, so `P(m == 0) = (1 - density)^lanes` crosses 0.5 inside
the useful density range for every width and the branch mispredicts
near-maximally right where it would pay off. Guarding cost 1.95x at density
0.5 for 4-byte elements and 1.45x at 0.1 for 1-byte. Storing a full chunk for
an empty mask is harmless because `write_pos` does not advance, so the next
store overwrites the garbage.

x86's `MIN_DENSITY_WIDE = 0.05` was set without AVX-512 hardware and with only
two density points measured, so it never probed its own crossover. On Zen 5 it
regressed 4-byte and 8-byte filtering by up to 32%, and AVX2 8-byte by 2.3x.
Both tiers now carry per-(kernel, width) floors, since what sets the floor is
how many lanes one compress covers.

Bands are measured per architecture, interleaved against the same binary with
the kernels disabled, and sit just above the crossover so a core with a
different shuffle-to-scalar balance cannot regress. Verified no regression at
any density on both architectures. Within band:

  NEON (Apple M-series):  1 byte 1.4-10.1x, 2 bytes 1.3-9.2x,
                          4 bytes 1.2-5.1x, 8 bytes 1.3x over 0.50..0.80
  x86 (Zen 5, AVX-512):   1 byte 1.9-5.7x, 2 bytes 1.1-3.9x,
                          4 bytes 1.1-2.8x, 8 bytes 1.1-3.1x
  x86 (Zen 5, AVX2 only): 4 bytes 1.1-2.5x, 8 bytes 1.4-2.0x

The 8-byte NEON kernel is the only one needing an upper bound: two lanes per
shuffle never amortizes far, and past ~0.8 the byte-LUT's all-set bulk copies
overtake it.

Also promotes simd_compress to a directory module, and installs mimalloc as
the global allocator in the benchmark. Without it the system allocator's
behaviour at these output sizes dominates and destabilizes divan's
per-iteration timing, swinging identical code by up to 20x.

Checks: cargo +nightly fmt --all --check; cargo clippy -p vortex-array
--all-targets --all-features; cargo nextest run -p vortex-array -p
vortex-buffer -p vortex-mask (4058 passed); cargo test --doc; rustdoc with
-D warnings. Full vortex-array suite also run on the AVX-512 machine (3180
passed), which is the first time the AVX-512 and AVX2 kernels have executed.
